Consolidate test suite: migrate to mirrored directories, remove duplicates
Migrated flat test files to 1:1 test-to-source directory structure,
merged split test files, and removed ~114 duplicate tests across 10 files.
