+++ tests/stages/test_coverage_gaps.py
@@ -18,118 +18,10 @@ _MOD = "azext_prototype.custom"
 class TestDeployHelpersDeep:
     """Deep tests for deploy_helpers module-level functions."""
 
-    # --- Bicep helpers ---
-
-    def test_find_bicep_params_json(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        (tmp_path / "main.parameters.json").write_text("{}", encoding="utf-8")
-        result = find_bicep_params(tmp_path, tmp_path / "main.bicep")
-        assert result is not None
-        assert result.name == "main.parameters.json"
-
-    def test_find_bicep_params_bicepparam(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        (tmp_path / "main.bicepparam").write_text("", encoding="utf-8")
-        result = find_bicep_params(tmp_path, tmp_path / "main.bicep")
-        assert result is not None
-        assert result.name == "main.bicepparam"
-
-    def test_find_bicep_params_generic(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        (tmp_path / "parameters.json").write_text("{}", encoding="utf-8")
-        result = find_bicep_params(tmp_path, tmp_path / "main.bicep")
-        assert result is not None
-        assert result.name == "parameters.json"
-
-    def test_find_bicep_params_none(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        result = find_bicep_params(tmp_path, tmp_path / "main.bicep")
-        assert result is None
-
-    def test_is_subscription_scoped_true(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import is_subscription_scoped
-
-        bicep = tmp_path / "main.bicep"
-        bicep.write_text("targetScope = 'subscription'\n", encoding="utf-8")
-        assert is_subscription_scoped(bicep) is True
-
-    def test_is_subscription_scoped_false(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import is_subscription_scoped
-
-        bicep = tmp_path / "main.bicep"
-        bicep.write_text("resource rg 'Microsoft.Resources/resourceGroups@2023-07-01' = {}\n", encoding="utf-8")
-        assert is_subscription_scoped(bicep) is False
-
     def test_is_subscription_scoped_missing_file(self, tmp_path):
         from azext_prototype.stages.deploy_helpers import is_subscription_scoped
 
         assert is_subscription_scoped(tmp_path / "nope.bicep") is False
-
-    def test_get_deploy_location_from_params(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        params = {"parameters": {"location": {"value": "westus2"}}}
-        (tmp_path / "parameters.json").write_text(json.dumps(params), encoding="utf-8")
-        assert get_deploy_location(tmp_path) == "westus2"
-
-    def test_get_deploy_location_from_string(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        params = {"location": "centralus"}
-        (tmp_path / "parameters.json").write_text(json.dumps(params), encoding="utf-8")
-        assert get_deploy_location(tmp_path) == "centralus"
-
-    def test_get_deploy_location_none(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        assert get_deploy_location(tmp_path) is None
-
-    def test_get_deploy_location_invalid_json(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        (tmp_path / "parameters.json").write_text("not json", encoding="utf-8")
-        assert get_deploy_location(tmp_path) is None
-
-    # --- check_az_login ---
-
-    @patch("azext_prototype.stages.deploy_helpers.subprocess.run")
-    def test_check_az_login_true(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import check_az_login
-
-        mock_run.return_value = MagicMock(returncode=0)
-        assert check_az_login() is True
-
-    @patch("azext_prototype.stages.deploy_helpers.subprocess.run")
-    def test_check_az_login_false(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import check_az_login
-
-        mock_run.return_value = MagicMock(returncode=1)
-        assert check_az_login() is False
-
-    @patch("azext_prototype.stages.deploy_helpers.subprocess.run", side_effect=FileNotFoundError)
-    def test_check_az_login_no_az(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import check_az_login
-
-        assert check_az_login() is False
-
-    # --- get_current_subscription ---
-
-    @patch("azext_prototype.stages.deploy_helpers.subprocess.run")
-    def test_get_current_subscription(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import get_current_subscription
-
-        mock_run.return_value = MagicMock(returncode=0, stdout="sub-abc-123\n")
-        assert get_current_subscription() == "sub-abc-123"
-
-    @patch("azext_prototype.stages.deploy_helpers.subprocess.run", side_effect=FileNotFoundError)
-    def test_get_current_subscription_error(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import get_current_subscription
-
-        assert get_current_subscription() == ""
 
     # --- deploy_terraform ---
