+++ tests/stages/test_deploy_session.py
@@ -550,42 +550,6 @@ class TestCaptureStageOutputs:
         session._output_capture.capture_bicep.assert_not_called()
 
 
-# ------------------------------------------------------------------
-# _extract_providers_from_files
-# ------------------------------------------------------------------
-
-
-class TestExtractProviders:
-    @patch("azext_prototype.stages.deploy_session._lookup_deployer_object_id", return_value=None)
-    @patch("azext_prototype.stages.deploy_session.build_deploy_env", return_value={})
-    @patch("azext_prototype.stages.deploy_session.get_current_subscription", return_value="sub")
-    def test_extracts_terraform_providers(self, mock_sub, mock_env, mock_oid, deploy_context, deploy_registry):
-        session = _make_session(deploy_context, deploy_registry)
-
-        stage_dir = Path(deploy_context.project_dir) / "concept" / "infra" / "terraform" / "stage-1-foundation"
-        stage_dir.mkdir(parents=True, exist_ok=True)
-        (stage_dir / "main.tf").write_text(
-            'resource "azapi_resource" "kv" {\n' '  type = "Microsoft.KeyVault/vaults@2023-07-01"\n' "}\n",
-            encoding="utf-8",
-        )
-
-        session._deploy_state._state["deployment_stages"] = [
-            {"stage": 1, "dir": "concept/infra/terraform/stage-1-foundation", "services": []}
-        ]
-
-        namespaces = session._extract_providers_from_files()
-        assert "Microsoft.KeyVault" in namespaces
-
-    @patch("azext_prototype.stages.deploy_session._lookup_deployer_object_id", return_value=None)
-    @patch("azext_prototype.stages.deploy_session.build_deploy_env", return_value={})
-    @patch("azext_prototype.stages.deploy_session.get_current_subscription", return_value="sub")
-    def test_no_files_returns_empty(self, mock_sub, mock_env, mock_oid, deploy_context, deploy_registry):
-        session = _make_session(deploy_context, deploy_registry)
-        session._deploy_state._state["deployment_stages"] = []
-        namespaces = session._extract_providers_from_files()
-        assert namespaces == set()
-
-
 # --- Additional imports from merged flat test ---
 from azext_prototype.agents.base import AgentContext
 from azext_prototype.agents.builtin import register_all_builtin
@@ -696,274 +660,6 @@ def _write_build_yaml(project_dir, stages=None, iac_tool="terraform"):
     with open(state_dir / "build.yaml", "w", encoding="utf-8") as f:
         yaml.dump(build_data, f, default_flow_style=False)
     return state_dir / "build.yaml"
-
-# ======================================================================
-
-
-class TestDeployState:
-
-    def test_default_state_structure(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        ds = DeployState(str(tmp_project))
-        state = ds.state
-        assert state["iac_tool"] == "terraform"
-        assert state["subscription"] == ""
-        assert state["resource_group"] == ""
-        assert state["deployment_stages"] == []
-        assert state["preflight_results"] == []
-        assert state["deploy_log"] == []
-        assert state["rollback_log"] == []
-        assert state["captured_outputs"] == {}
-        assert state["_metadata"]["iteration"] == 0
-
-    def test_load_save_roundtrip(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        ds = DeployState(str(tmp_project))
-        ds._state["subscription"] = "test-sub-123"
-        ds._state["iac_tool"] = "bicep"
-        ds.save()
-
-        ds2 = DeployState(str(tmp_project))
-        loaded = ds2.load()
-        assert loaded["subscription"] == "test-sub-123"
-        assert loaded["iac_tool"] == "bicep"
-        assert loaded["_metadata"]["created"] is not None
-        assert loaded["_metadata"]["last_updated"] is not None
-
-    def test_exists_property(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        ds = DeployState(str(tmp_project))
-        assert not ds.exists
-        ds.save()
-        assert ds.exists
-
-    def test_load_from_build_state(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        result = ds.load_from_build_state(build_path)
-
-        assert result is True
-        assert len(ds.state["deployment_stages"]) == 3
-        # Verify deploy-specific fields were added
-        stage = ds.state["deployment_stages"][0]
-        assert stage["deploy_status"] == "pending"
-        assert stage["deploy_timestamp"] is None
-        assert stage["deploy_output"] == ""
-        assert stage["deploy_error"] == ""
-        assert stage["rollback_timestamp"] is None
-
-    def test_load_from_build_state_missing_file(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        ds = DeployState(str(tmp_project))
-        result = ds.load_from_build_state("/nonexistent/build.yaml")
-        assert result is False
-
-    def test_load_from_build_state_no_stages(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project, stages=[])
-        ds = DeployState(str(tmp_project))
-        result = ds.load_from_build_state(build_path)
-        assert result is False
-
-    def test_stage_transitions(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-
-        # pending → deploying
-        ds.mark_stage_deploying(1)
-        assert ds.get_stage(1)["deploy_status"] == "deploying"
-
-        # deploying → deployed
-        ds.mark_stage_deployed(1, output="resource_id=abc123")
-        stage = ds.get_stage(1)
-        assert stage["deploy_status"] == "deployed"
-        assert stage["deploy_timestamp"] is not None
-        assert stage["deploy_output"] == "resource_id=abc123"
-        assert stage["deploy_error"] == ""
-
-        # deployed → rolled_back
-        ds.mark_stage_rolled_back(1)
-        stage = ds.get_stage(1)
-        assert stage["deploy_status"] == "rolled_back"
-        assert stage["rollback_timestamp"] is not None
-
-    def test_stage_failure(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-
-        ds.mark_stage_deploying(1)
-        ds.mark_stage_failed(1, error="timeout connecting to Azure")
-        stage = ds.get_stage(1)
-        assert stage["deploy_status"] == "failed"
-        assert stage["deploy_error"] == "timeout connecting to Azure"
-
-    def test_get_pending_deployed_failed(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-
-        assert len(ds.get_pending_stages()) == 3
-        assert len(ds.get_deployed_stages()) == 0
-        assert len(ds.get_failed_stages()) == 0
-
-        ds.mark_stage_deployed(1)
-        ds.mark_stage_failed(2, "error")
-
-        assert len(ds.get_pending_stages()) == 1
-        assert len(ds.get_deployed_stages()) == 1
-        assert len(ds.get_failed_stages()) == 1
-
-    def test_can_rollback_ordering(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-
-        ds.mark_stage_deployed(1)
-        ds.mark_stage_deployed(2)
-        ds.mark_stage_deployed(3)
-
-        # Can only rollback stage 3 (highest)
-        assert ds.can_rollback(3) is True
-        assert ds.can_rollback(2) is False  # stage 3 still deployed
-        assert ds.can_rollback(1) is False  # stages 2,3 still deployed
-
-        # Roll back stage 3
-        ds.mark_stage_rolled_back(3)
-        assert ds.can_rollback(2) is True
-        assert ds.can_rollback(1) is False  # stage 2 still deployed
-
-    def test_rollback_candidates_reverse_order(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-
-        ds.mark_stage_deployed(1)
-        ds.mark_stage_deployed(2)
-        ds.mark_stage_deployed(3)
-
-        candidates = ds.get_rollback_candidates()
-        assert [c["stage"] for c in candidates] == [3, 2, 1]
-
-    def test_preflight_results(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        ds = DeployState(str(tmp_project))
-        results = [
-            {"name": "Azure Login", "status": "pass", "message": "Logged in."},
-            {"name": "Terraform", "status": "fail", "message": "Not found.", "fix_command": "brew install terraform"},
-        ]
-        ds.set_preflight_results(results)
-
-        failures = ds.get_preflight_failures()
-        assert len(failures) == 1
-        assert failures[0]["name"] == "Terraform"
-
-    def test_deploy_log(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-
-        ds.mark_stage_deploying(1)
-        ds.mark_stage_deployed(1)
-
-        assert len(ds.state["deploy_log"]) == 2
-        assert ds.state["deploy_log"][0]["action"] == "deploying"
-        assert ds.state["deploy_log"][1]["action"] == "deployed"
-
-    def test_reset(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-        assert len(ds.state["deployment_stages"]) == 3
-
-        ds.reset()
-        assert ds.state["deployment_stages"] == []
-        assert ds.exists  # File still exists after reset
-
-    def test_format_deploy_report(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-        ds._state["subscription"] = "sub-123"
-
-        ds.mark_stage_deployed(1)
-        ds.mark_stage_failed(2, "timeout")
-
-        report = ds.format_deploy_report()
-        assert "Deploy Report" in report
-        assert "sub-123" in report
-        assert "1 deployed" in report
-        assert "1 failed" in report
-
-    def test_format_stage_status(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        build_path = _write_build_yaml(tmp_project)
-        ds = DeployState(str(tmp_project))
-        ds.load_from_build_state(build_path)
-
-        status = ds.format_stage_status()
-        assert "Foundation" in status
-        assert "Application" in status
-        assert "0/3 stages deployed" in status
-
-    def test_format_preflight_report(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        ds = DeployState(str(tmp_project))
-        ds.set_preflight_results(
-            [
-                {"name": "Azure Login", "status": "pass", "message": "OK"},
-                {
-                    "name": "Terraform",
-                    "status": "warn",
-                    "message": "Old version",
-                    "fix_command": "brew upgrade terraform",
-                },
-            ]
-        )
-
-        report = ds.format_preflight_report()
-        assert "Preflight Checks" in report
-        assert "2 passed" in report or "1 passed" in report
-        assert "1 warning" in report
-
-    def test_conversation_tracking(self, tmp_project):
-        from azext_prototype.stages.deploy_state import DeployState
-
-        ds = DeployState(str(tmp_project))
-        ds.update_from_exchange("deploy all", "Deploying stage 1...", 1)
-
-        assert len(ds.state["conversation_history"]) == 1
-        assert ds.state["conversation_history"][0]["user"] == "deploy all"
-
-# ======================================================================
-
 
 class TestExtractResourceProvidersFromFiles:
     """Verify _extract_providers_from_files() parses IaC files for namespaces."""
