+++ tests/stages/test_stages.py
@@ -179,211 +179,9 @@ class TestInitStage:
         assert len(guards) == 0
 
 
-class TestDeployStage:
-    """Test the deploy stage."""
-
-    def test_deploy_stage_instantiates(self):
-        from azext_prototype.stages.deploy_stage import DeployStage
-
-        stage = DeployStage()
-        assert stage is not None
-        assert stage.name == "deploy"
-
-    def test_deploy_stage_has_execute(self):
-        from azext_prototype.stages.deploy_stage import DeployStage
-
-        stage = DeployStage()
-        assert hasattr(stage, "execute")
-        assert callable(stage.execute)
-
-
-class TestDeployBicepStaging:
-    """Test Bicep staged deployment capabilities (via deploy_helpers)."""
-
-    def test_find_bicep_params_main_parameters_json(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        template = tmp_path / "main.bicep"
-        template.write_text("resource rg 'Microsoft.Resources/resourceGroups@2023-07-01' = {}")
-        params = tmp_path / "main.parameters.json"
-        params.write_text('{"parameters": {"location": {"value": "eastus"}}}')
-
-        result = find_bicep_params(tmp_path, template)
-        assert result == params
-
-    def test_find_bicep_params_bicepparam(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        template = tmp_path / "main.bicep"
-        template.write_text("")
-        bp = tmp_path / "main.bicepparam"
-        bp.write_text("using './main.bicep'\nparam location = 'eastus'")
-
-        result = find_bicep_params(tmp_path, template)
-        assert result == bp
-
-    def test_find_bicep_params_fallback_parameters_json(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        template = tmp_path / "network.bicep"
-        template.write_text("")
-        params = tmp_path / "parameters.json"
-        params.write_text('{"parameters": {}}')
-
-        result = find_bicep_params(tmp_path, template)
-        assert result == params
-
-    def test_find_bicep_params_none_when_missing(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import find_bicep_params
-
-        template = tmp_path / "main.bicep"
-        template.write_text("")
-
-        result = find_bicep_params(tmp_path, template)
-        assert result is None
-
-    def test_is_subscription_scoped_true(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import is_subscription_scoped
-
-        bicep_file = tmp_path / "main.bicep"
-        bicep_file.write_text(
-            "targetScope = 'subscription'\n\nresource rg 'Microsoft.Resources/resourceGroups@2023-07-01' = {}"
-        )
-
-        assert is_subscription_scoped(bicep_file) is True
-
-    def test_is_subscription_scoped_false(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import is_subscription_scoped
-
-        bicep_file = tmp_path / "main.bicep"
-        bicep_file.write_text("resource sa 'Microsoft.Storage/storageAccounts@2023-01-01' = {}")
-
-        assert is_subscription_scoped(bicep_file) is False
-
-    def test_get_deploy_location_from_params(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        params = tmp_path / "parameters.json"
-        params.write_text('{"parameters": {"location": {"value": "westus2"}}}')
-
-        result = get_deploy_location(tmp_path)
-        assert result == "westus2"
-
-    def test_get_deploy_location_returns_none(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        result = get_deploy_location(tmp_path)
-        assert result is None
-
-    @patch("subprocess.run")
-    def test_deploy_bicep_resource_group_scope(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_bicep
-
-        bicep_dir = tmp_path / "stage1"
-        bicep_dir.mkdir()
-        (bicep_dir / "main.bicep").write_text("resource sa 'Microsoft.Storage/storageAccounts@2023-01-01' = {}")
-
-        mock_run.return_value = MagicMock(returncode=0, stdout='{"properties":{}}', stderr="")
-
-        result = deploy_bicep(bicep_dir, "sub-123", "my-rg")
-        assert result["status"] == "deployed"
-        assert result["scope"] == "resourceGroup"
-        assert result["template"] == "main.bicep"
-
-        # Verify az deployment group create was called (not sub create)
-        cmd_parts = mock_run.call_args[0][0]
-        assert "group" in cmd_parts
-        assert "create" in cmd_parts
-
-    @patch("subprocess.run")
-    def test_deploy_bicep_subscription_scope(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_bicep
-
-        bicep_dir = tmp_path / "stage1"
-        bicep_dir.mkdir()
-        (bicep_dir / "main.bicep").write_text(
-            "targetScope = 'subscription'\n\nresource rg 'Microsoft.Resources/resourceGroups@2023-07-01' = {}"
-        )
-
-        mock_run.return_value = MagicMock(returncode=0, stdout='{"properties":{}}', stderr="")
-
-        result = deploy_bicep(bicep_dir, "sub-123", "")
-        assert result["status"] == "deployed"
-        assert result["scope"] == "subscription"
-
-        # Verify az deployment sub create was called
-        cmd_parts = mock_run.call_args[0][0]
-        assert "sub" in cmd_parts
-
-    @patch("subprocess.run")
-    def test_deploy_bicep_with_params_file(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_bicep
-
-        (tmp_path / "main.bicep").write_text("param location string\n")
-        (tmp_path / "main.parameters.json").write_text('{"parameters":{"location":{"value":"eastus"}}}')
-
-        mock_run.return_value = MagicMock(returncode=0, stdout="{}", stderr="")
-
-        deploy_bicep(tmp_path, "sub-123", "my-rg")
-
-        cmd_parts = mock_run.call_args[0][0]
-        assert "--parameters" in cmd_parts
-
-    def test_deploy_bicep_no_bicep_files_skips(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_bicep
-
-        empty_dir = tmp_path / "empty"
-        empty_dir.mkdir()
-
-        result = deploy_bicep(empty_dir, "sub-123", "my-rg")
-        assert result["status"] == "skipped"
-
-    def test_deploy_bicep_fallback_to_first_file(self, tmp_path):
-        """When no main.bicep exists, uses the first .bicep file."""
-        from azext_prototype.stages.deploy_helpers import deploy_bicep
-
-        (tmp_path / "network.bicep").write_text("resource vnet 'Microsoft.Network/virtualNetworks@2023-05-01' = {}")
-
-        with patch("subprocess.run") as mock_run:
-            mock_run.return_value = MagicMock(returncode=0, stdout="{}", stderr="")
-            result = deploy_bicep(tmp_path, "sub-123", "my-rg")
-
-        assert result["status"] == "deployed"
-        assert result["template"] == "network.bicep"
-
-    def test_deploy_bicep_rg_required_for_rg_scope(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_bicep
-
-        (tmp_path / "main.bicep").write_text("resource sa 'Microsoft.Storage/storageAccounts@2023-01-01' = {}")
-
-        result = deploy_bicep(tmp_path, "sub-123", "")
-        assert result["status"] == "failed"
-        assert "Resource group required" in result["error"]
-
-    @patch("subprocess.run")
-    def test_whatif_bicep_runs(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import whatif_bicep
-
-        (tmp_path / "main.bicep").write_text("resource sa 'Microsoft.Storage/storageAccounts@2023-01-01' = {}")
-        mock_run.return_value = MagicMock(returncode=0, stdout="Resource changes: 1 to create", stderr="")
-
-        result = whatif_bicep(tmp_path, "sub-123", "my-rg")
-        assert result["status"] == "previewed"
-        assert "Resource changes" in result["output"]
-
-        cmd_parts = mock_run.call_args[0][0]
-        assert "what-if" in cmd_parts
-
 
 class TestDesignStage:
     """Test the design stage."""
-
-    def test_design_stage_instantiates(self):
-        from azext_prototype.stages.design_stage import DesignStage
-
-        stage = DesignStage()
-        assert stage.name == "design"
-        assert stage.reentrant is True
 
     def test_design_stage_has_execute(self):
         from azext_prototype.stages.design_stage import DesignStage
@@ -1021,25 +819,6 @@ class TestDesignStage:
             )
 
 
-class TestBuildStage:
-    """Test the build stage."""
-
-    def test_build_stage_instantiates(self):
-        from azext_prototype.stages.build_stage import BuildStage
-
-        stage = BuildStage()
-        assert stage is not None
-        assert stage.name == "build"
-
-    def test_match_templates_empty_architecture(self):
-        from azext_prototype.stages.build_stage import BuildStage
-
-        stage = BuildStage()
-        config = MagicMock()
-        result = stage._match_templates({"architecture": ""}, config)
-        assert result == []
-
-
 # --- Additional imports from merged flat test ---
 from knack.util import CLIError
 
@@ -1047,190 +826,7 @@ from azext_prototype.agents.base import AgentContext
 from azext_prototype.agents.registry import AgentRegistry
 from azext_prototype.ai.provider import AIResponse
 from azext_prototype.stages.build_session import BuildResult
-from azext_prototype.stages.deploy_helpers import check_az_login
-from azext_prototype.stages.deploy_helpers import deploy_app_stage
-from azext_prototype.stages.deploy_helpers import deploy_terraform
-from azext_prototype.stages.deploy_helpers import get_current_subscription
 
-
-# ======================================================================
-
-
-class TestDeployStageExecution:
-    """Test DeployStage orchestration and deploy_helpers functions."""
-
-    def _make_stage(self):
-        from azext_prototype.stages.deploy_stage import DeployStage
-
-        return DeployStage()
-
-    def test_deploy_guards(self):
-        stage = self._make_stage()
-        guards = stage.get_guards()
-        names = [g.name for g in guards]
-        assert "project_initialized" in names
-        assert "build_complete" in names
-        assert "az_logged_in" in names
-
-    @patch("subprocess.run")
-    def test_check_az_login_true(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import check_az_login
-
-        mock_run.return_value = MagicMock(returncode=0)
-        assert check_az_login() is True
-
-    @patch("subprocess.run")
-    def test_check_az_login_false(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import check_az_login
-
-        mock_run.return_value = MagicMock(returncode=1)
-        assert check_az_login() is False
-
-    @patch("subprocess.run", side_effect=FileNotFoundError)
-    def test_check_az_login_not_installed(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import check_az_login
-
-        assert check_az_login() is False
-
-    @patch("subprocess.run")
-    def test_get_current_subscription(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import get_current_subscription
-
-        mock_run.return_value = MagicMock(returncode=0, stdout="abc-123\n")
-        result = get_current_subscription()
-        assert result == "abc-123"
-
-    @patch("subprocess.run", side_effect=FileNotFoundError)
-    def test_get_current_subscription_not_installed(self, mock_run):
-        from azext_prototype.stages.deploy_helpers import get_current_subscription
-
-        assert get_current_subscription() == ""
-
-    @patch("subprocess.run")
-    def test_deploy_terraform_success(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_terraform
-
-        infra_dir = tmp_path / "tf"
-        infra_dir.mkdir()
-        mock_run.return_value = MagicMock(returncode=0, stdout="", stderr="")
-
-        result = deploy_terraform(infra_dir, "sub-123")
-        assert result["status"] == "deployed"
-
-    @patch("subprocess.run")
-    def test_deploy_terraform_failure(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_terraform
-
-        infra_dir = tmp_path / "tf"
-        infra_dir.mkdir()
-        mock_run.return_value = MagicMock(returncode=1, stderr="init failed", stdout="")
-
-        result = deploy_terraform(infra_dir, "sub-123")
-        assert result["status"] == "failed"
-
-    @patch("subprocess.run")
-    def test_deploy_bicep_failure(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_bicep
-
-        (tmp_path / "main.bicep").write_text("resource x 'y' = {}", encoding="utf-8")
-        mock_run.return_value = MagicMock(returncode=1, stderr="Deployment failed", stdout="")
-
-        result = deploy_bicep(tmp_path, "sub-123", "my-rg")
-        assert result["status"] == "failed"
-
-    def test_deploy_app_stage_with_deploy_script(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_app_stage
-
-        app_dir = tmp_path / "app"
-        app_dir.mkdir()
-        (app_dir / "deploy.sh").write_text("echo deployed", encoding="utf-8")
-
-        with patch("subprocess.run") as mock_run:
-            mock_run.return_value = MagicMock(returncode=0, stdout="", stderr="")
-            result = deploy_app_stage(app_dir, "sub-123", "my-rg")
-            assert result["status"] == "deployed"
-
-    def test_deploy_app_stage_sub_apps(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_app_stage
-
-        stage_dir = tmp_path / "stage"
-        stage_dir.mkdir()
-        backend = stage_dir / "backend"
-        backend.mkdir()
-        (backend / "deploy.sh").write_text("echo ok", encoding="utf-8")
-
-        with patch("subprocess.run") as mock_run:
-            mock_run.return_value = MagicMock(returncode=0, stdout="", stderr="")
-            result = deploy_app_stage(stage_dir, "sub-123", "my-rg")
-            assert result["status"] == "deployed"
-            assert "backend" in result["apps"]
-
-    def test_deploy_app_stage_no_scripts(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import deploy_app_stage
-
-        empty_dir = tmp_path / "empty"
-        empty_dir.mkdir()
-        result = deploy_app_stage(empty_dir, "sub-123", "my-rg")
-        assert result["status"] == "skipped"
-
-    @patch("subprocess.run")
-    def test_whatif_bicep_no_files(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import whatif_bicep
-
-        empty_dir = tmp_path / "empty"
-        empty_dir.mkdir()
-        result = whatif_bicep(empty_dir, "sub-123", "my-rg")
-        assert result["status"] == "skipped"
-
-    @patch("subprocess.run")
-    def test_whatif_bicep_no_rg_skips(self, mock_run, tmp_path):
-        from azext_prototype.stages.deploy_helpers import whatif_bicep
-
-        (tmp_path / "main.bicep").write_text("resource x 'y' = {}", encoding="utf-8")
-        result = whatif_bicep(tmp_path, "sub-123", "")
-        assert result["status"] == "skipped"
-
-    def test_get_deploy_location_main_params(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        (tmp_path / "main.parameters.json").write_text(
-            '{"parameters": {"location": {"value": "northeurope"}}}', encoding="utf-8"
-        )
-        result = get_deploy_location(tmp_path)
-        assert result == "northeurope"
-
-    def test_get_deploy_location_string_value(self, tmp_path):
-        from azext_prototype.stages.deploy_helpers import get_deploy_location
-
-        (tmp_path / "parameters.json").write_text('{"location": "uksouth"}', encoding="utf-8")
-        result = get_deploy_location(tmp_path)
-        assert result == "uksouth"
-
-    def test_execute_status(self, project_with_build, mock_agent_context, populated_registry):
-        """Deploy with --status shows state and returns."""
-        stage = self._make_stage()
-        stage.get_guards = lambda: []
-        mock_agent_context.project_dir = str(project_with_build)
-
-        result = stage.execute(
-            mock_agent_context,
-            populated_registry,
-            status=True,
-        )
-        assert result["status"] == "status_displayed"
-
-    def test_execute_reset(self, project_with_build, mock_agent_context, populated_registry):
-        """Deploy with --reset clears state and returns."""
-        stage = self._make_stage()
-        stage.get_guards = lambda: []
-        mock_agent_context.project_dir = str(project_with_build)
-
-        result = stage.execute(
-            mock_agent_context,
-            populated_registry,
-            reset=True,
-        )
-        assert result["status"] == "reset"
 
 # ======================================================================
 
@@ -1242,32 +838,6 @@ class TestBuildStageExecution:
         from azext_prototype.stages.build_stage import BuildStage
 
         return BuildStage()
-
-    def test_build_guards(self):
-        stage = self._make_stage()
-        guards = stage.get_guards()
-        names = [g.name for g in guards]
-        assert "project_initialized" in names
-        assert "discovery_complete" in names
-        assert "design_complete" in names
-
-    def test_load_design(self, project_with_design):
-        stage = self._make_stage()
-        design = stage._load_design(str(project_with_design))
-        assert "architecture" in design
-
-    def test_load_design_missing(self, tmp_project):
-        stage = self._make_stage()
-        result = stage._load_design(str(tmp_project))
-        assert result == {}
-
-    def test_execute_no_design_raises(self, project_with_config, mock_agent_context, populated_registry):
-        stage = self._make_stage()
-        stage.get_guards = lambda: []
-        mock_agent_context.project_dir = str(project_with_config)
-
-        with pytest.raises(CLIError, match="No architecture design"):
-            stage.execute(mock_agent_context, populated_registry)
 
     def test_execute_dry_run(self, project_with_design, mock_agent_context, populated_registry):
         stage = self._make_stage()
