+++ tests/telemetry/test___init__.py
@@ -355,7 +355,7 @@ class TestGetExtensionVersion:
         from azext_prototype.telemetry import _get_extension_version
 
         version = _get_extension_version()
-        assert version == "0.2.1b6"
+        assert version == "0.2.1b7"
 
     def test_returns_unknown_on_error(self):
         from azext_prototype.telemetry import _get_extension_version
